requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.3
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
cx_Freeze==6.15.16
cx_Logging==3.2.0
cycler==0.11.0
darkdetect==0.8.0
decorator==5.1.1
dnspython==2.6.1
email_validator==2.1.1
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.110.0
Flask==2.3.2
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.42.0
frozenlist==1.3.3
google==3.0.0
google-api-core==2.18.0
google-api-python-client==2.124.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
gspread==6.1.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.4
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
kiwisolver==1.4.4
lief==0.14.1
MarkupSafe==2.1.3
mediapipe==0.10.3
multidict==6.0.4
networkx==3.2.1
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
orjson==3.10.0
packaging==23.1
parso==0.8.3
pefile==2023.2.7
pickleshare==0.7.5
plum-dispatch==1.7.4
prompt-toolkit==3.0.39
proto-plus==1.23.0
protobuf==3.20.3
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.6.4
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
Pygments==2.16.1
pynput==1.7.6
pyocr==0.8.5
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2023.3
PyYAML==6.0.1
requests==2.28.2
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.2
starlette==0.36.3
StrEnum==0.4.15
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.10.0
tzdata==2023.3
ujson==5.9.0
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.29.0
waitress==3.0.0
watchfiles==0.21.0
wcwidth==0.2.6
websockets==12.0
Werkzeug==2.3.7
xlrd==2.0.1
yarl==1.8.2
